Optimizer: convert the EscapeInfoDumper from a pass to a test
