+++ SwiftCompilerSources/Sources/Optimizer/TestPasses/CMakeLists.txt
@@ -11,7 +11,6 @@ swift_compiler_sources(Optimizer
   AccessDumper.swift
   AliasInfoDumper.swift
   DeadEndBlockDumper.swift
-  EscapeInfoDumper.swift
   MemBehaviorDumper.swift
   SILPrinter.swift
   RangeDumper.swift
